restaurant/requirements.txt
Django==2.2.5
django-admin-generator==2.2.0
django-admin-interface==0.11.2
django-filebrowser-no-grappelli==3.7.8
django-seed==0.1.9
djangorestframework==3.10.3
djangorestframework-api-key==1.4.1
djangorestframework-simplejwt==4.3.0
drf-dynamic-fields==0.3.1
drf-nested-routers==0.91
drf-yasg==1.17.0
Pillow==6.1.0
rest-framework-generic-relations